sites/urbanism-now-links/pyproject.toml
[project]
name = "urbanism-now-links"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "cachetools>=5.5.1",
    "notion-client>=2.3.0",
    "openai>=1.61.0",
    "python-dotenv>=1.0.1",
    "trafilatura>=2.0.0",
]

[dependency-groups]
dev = [
    "ruff>=0.9.4",
]

[tool.ruff.lint]
select = [
    "W",  # pywarn
    "I",  # isort
]
